+++ libraries/ZWidget/src/window/win32/win32_util.h
@@ -5,8 +5,8 @@
 #ifndef WINVER
 #define WINVER 0x0605
 #endif
-#include <Windows.h>
-#include <Shlobj.h>
+#include <windows.h>
+#include <shlobj.h>
 #include <stdexcept>
 
 namespace