+++ libraries/lzma/C/7zWindows.h
@@ -31,7 +31,7 @@
 // #if defined(__GNUC__) && !defined(__clang__)
 #include <windows.h>
 #else
-#include <Windows.h>
+#include <windows.h>
 #endif
 // #include <basetsd.h>
 // #include <wtypes.h>