+++ src/common/platform/win32/i_mainwindow.h
@@ -6,7 +6,7 @@
 #include <functional>
 
 #define WIN32_LEAN_AND_MEAN
-#include <Windows.h>
+#include <windows.h>
 
 // The WndProc used when the game view is active
 LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam);